Move composition provider to its own file
